Update checkGridForDestroy
Updated checkGridForDestroy to theoretically work faster. I assume the project is too simple to really cause any noticeable improvement, but this is still faster.

Specifically, uses a colorArray to check if there are 3+ bubbles in a row (as in next to each other) in both rows and columns. This does 10 checks rather than 18 (and I think mathematically there is a greater potential difference, but the 18->10 is what I care about).

Also, because the function now uses a color array I can update the bubbles' color when they are also marked for destruction rather than afterward which was wasteful.